requirements.txt
# 核心依赖
playwright==1.40.0
playwright-stealth==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Streamlit
streamlit==1.28.0

# 工具
python-dotenv==1.0.0
pydantic==2.5.0
